+++ changelog.md
@@ -11,6 +11,7 @@
 - Get display name from WW on login
 - width for extra rounds form
 - show assets field error on series create
+- If series recording access is GUILD any guild members can be added to game
 
 ## [1.12.10] - 2026-6-30 Small Bugs Need Squashing Too